Adding the tests that somehow didn't get added in last commit; ensured removal of original test class
