--- Assets/PlayTest/CreatureBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0e1148d0c0f0d5d4cbec4c1d74e139c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
